requirements.txt
python-dotenv
google-generativeai
PyPDF2
python-docx
Pillow
pytesseract
streamlit
fpdf==1.7.2
